server: Use zaptest in bucket tests and move backendMock to separate file
