+++ server/storage/schema/auth_roles_test.go
@@ -19,7 +19,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"go.uber.org/zap"
+	"go.uber.org/zap/zaptest"
 
 	"go.etcd.io/etcd/api/v3/authpb"
 	"go.etcd.io/etcd/server/v3/auth"
@@ -110,7 +110,7 @@ func TestGetAllRoles(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(t *testing.T) {
 			be, tmpPath := betesting.NewTmpBackend(t, time.Microsecond, 10)
-			abe := NewAuthBackend(zap.NewNop(), be)
+			abe := NewAuthBackend(zaptest.NewLogger(t), be)
 			abe.CreateAuthBuckets()
 
 			tx := abe.BatchTx()
@@ -123,7 +123,7 @@ func TestGetAllRoles(t *testing.T) {
 
 			be2 := backend.NewDefaultBackend(tmpPath)
 			defer be2.Close()
-			abe2 := NewAuthBackend(zap.NewNop(), be2)
+			abe2 := NewAuthBackend(zaptest.NewLogger(t), be2)
 			users := abe2.GetAllRoles()
 
 			assert.Equal(t, tc.want, users)
@@ -206,7 +206,7 @@ func TestGetRole(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(t *testing.T) {
 			be, tmpPath := betesting.NewTmpBackend(t, time.Microsecond, 10)
-			abe := NewAuthBackend(zap.NewNop(), be)
+			abe := NewAuthBackend(zaptest.NewLogger(t), be)
 			abe.CreateAuthBuckets()
 
 			tx := abe.BatchTx()
@@ -219,7 +219,7 @@ func TestGetRole(t *testing.T) {
 
 			be2 := backend.NewDefaultBackend(tmpPath)
 			defer be2.Close()
-			abe2 := NewAuthBackend(zap.NewNop(), be2)
+			abe2 := NewAuthBackend(zaptest.NewLogger(t), be2)
 			users := abe2.GetRole("role1")
 
 			assert.Equal(t, tc.want, users)
